--HG--
branch : DAGuE

+++ data_dist/matrix/matrix.h
@@ -66,7 +66,7 @@ void generate_tiled_zero_mat(tiled_matrix_desc_t * Mdesc);
 /**
  * allocate a buffer to hold the matrix
  */
-void* dague_allocate_matrix( int matrix_size );
+void* dague_allocate_matrix(uint64_t matrix_size );
 
 int data_write(tiled_matrix_desc_t * Ddesc, char * filename);
